Syncing with FreeImage repo, readme update

+++ src/freeimage/FreeImage/PluginTIFF.cpp
@@ -204,7 +204,7 @@ TIFFFdOpen(thandle_t handle, const char *name, const char *mode) {
     // 64bit machines (sizeof(int) != sizeof(long)). 
     // Needs to be fixed within libTIFF.
 	if (tif) {
-		tif->tif_fd = (long)handle;
+		tif->tif_fd = (intptr_t)handle;
 	}
 
 	return tif;
